==> 06-references.Rmd <==
#References {#references}

1. Bruce, D. F. (2009). __Asthma and Depression__. _WedMD_ (2009). Retrieved from https://www.webmd.com/asthma/features/asthma-depression in 12/1/2018

2. Mayo Clinic. (2018). __Asthma__. _Mayo Clinic_ (2018). Retrieved from  https://www.mayoclinic.org/diseases-conditions/asthma/symptoms-causes/syc-20369653 in 12/1/2018

3. National Center for Health Statistics (2012). __Trends in Asthma Prevalence, Health Care Use, and Mortality in the United States, 2001 - 2010.__ _Centers for Disease Control and Prevention_ (2012) Retrieved from  https://www.cdc.gov/nchs/data/databriefs/db94.htm in 12/1/2018
